refactor: don't globally expose userscript metadata types
